Reorganize and document storage adapters
